requirements.txt
streamlit
requests
python-dotenv
fastapi
uvicorn
langchain
google-generativeai
langchain-google-genai